+++ src/octopus/src/controller/Controller.hh
@@ -82,7 +82,7 @@ class Controller
 
 public:
 	Controller(std::list<Steppable *> const &initSteppables_p, double timePerStep_p,
-		std::list<Command *> const &initCommands_p = {}, unsigned long gridPointSize_p=1, unsigned long gridSize_p=50, size_t reusableHandleQueueSize_p=500);
+		std::list<Command *> const &initCommands_p = {}, unsigned long gridPointSize_p=1, unsigned long gridSize_p=50, uint32_t reusableHandleQueueSize_p=500);
 	~Controller();
 
 	/// @brief returns true once buffer and back states are up to date
@@ -118,7 +118,7 @@ public:
 	void addQueuedLayer();
 
 	/// @brief get the queued layer command size
-	size_t getQueuedSize() const;
+	uint32_t getQueuedSize() const;
 
 	/// @brief add a trigger on the ongoing step
 	void commitTrigger(Trigger * trigger_p);
@@ -228,7 +228,7 @@ private:
 
 	/// @brief the size of the queue in states for reusable handles
 	/// (this means a free handle wont be reusable before _reusablehandleQueueSize steps have been applied)
-	size_t const _reusableHandleQueueSize;
+	uint32_t const _reusableHandleQueueSize;
 };
 
 } // namespace octopus
