+++ src/octopus/src/orca/rvo/Obstacle.hh
@@ -57,7 +57,7 @@ namespace RVO {
 		Obstacle *prevObstacle_;
 		Vector2 unitDir_;
 
-		size_t id_;
+		uint32_t id_;
 
 		friend class Agent;
 		friend class KdTree;
